Generate required rules for notNull columns.
Historically this was not that important because with type-safe
EntityManager.create and entity constructors, it was surprisingly
hard to create entities with missing required fields.

However with createUnsafe and setUnsafe for dealing with more
real-world/partial-update-ish/etc. APIs, it's more likely for
required values to sneak in as being unset.