Optimize findExistingInstance and getMetadata.
Profiling a query that loaded ~2,000 entities showed large hot-spots
in `findExistingInstance` + `isEntity` + `getMetadata`. It was taking
~60 seconds and eventually timing out.

Adding an index for `findExistingInstance` dropped the time for the
same query to ~3 seconds and then it fails with a "too many entities"
error.

Kinda surprised at how much of a difference this makes, but when loading
~1000s of entities, findExistingInstance is called in a tight loop and
turns out O(N^2) is not a great idea.