Add incremental collection support via op keyword.
Our collections used to be incremental-by-default, but we removed
that b/c it was unintuitive / violated least-surprise to send in
`collection: [1, 2]` and have that leave an existing `3` in the
collection.

So we changed the behavior to exhaustive only.

This adds back opting in to incremental behavior, if all children
in a given `set` operation have an `op` keyword that, for each
child, specifies whether to include/remove/delete it.