Better handle children creating parents via factories. (#145)
I.e. if a parent factory has:

```
export function newParent(...) {
  return newTestInstance(em, {
    // Always have one child
    children: [{}],
    ...opts
  });
}
```

But a test is creating the child "bottom up", i.e. doing:

```
newChild(em, { parent: ... });
```

Then we know the parent is going to get a list of children already (from
the one that is currently being created), so we don't ned its `children:
[]` default to apply.

So we send in `opts.children: null`, as if the programmer had written
that by hand, such that the `...opts` convention of the factories will
write over the `children: [{}]` default.