feat: Make clone smarter.
Previously clone would both find + fixup entities in a single pass.

This made it hard/impossible for clone to go back and update
already-cloned entities to point to new downstream clones.

I.e. when cloning author -> book, author has a currentDraftBook field,
but we've not even cloned it's books yet, so would not fix up the
relation, and leave a newly cloned author pointing to the original
author's book.

This PR breaks up clone into three distinct phases of 1) find, 2)
clone, and 3) fixup.

The upshot is that the fixup phase now knows about all clones, i.e.
we can point `author.currentDraftBook` to the new book.