fix: Fix async derived fields not triggering from hook changes.
We need to run the hooks first, to see if they have any side effects,
and then detect which rules & async derived fields we need re-calc.

...granted, this begs the question of what about hooks that want to
observe the latest derived value? This is probably where we need a
DAG of hook/value dependencies and invoke them in topo order instead
of fixed.