fix: Remove id from allowed factory params.
This seems cute at the time, but in practice we basically never use
it, and always use variables or hashes.

And keeping it in throws off intellisense b/c doing things like
`author: { ... }` within the `...` you'll be prompted for string-ish
methods, like if you had wanted to pass a param that structurally
matches string.

So net/net better DX is to only see the entities fields and not the
misc string methods in the autocomplete list.