fix: Fix derived fields on subtypes.
The derived field on the subtype itself worked, but entities of the
_other_ subtype would blow up.

For example, `SmallPublisher.allAuthorNames` is a derived field.

If an Author points to a SmallPublisher, we can follow a1 -> publishers
-> recalc sp1[allAuthorNames].

But if an Author points to a LargePublisher, if we try to follow a1 ->
publishers -> recalc lp1[allAuthorNames], it will NPE because the
LargePublisher doesn't have an `allAuthorNames` field.

Which is fine, the Author, when it said "hey tell my publisher to recalc
things" can't know "is this publisher large or small?", so we just have
to detect/infer that at runtime, by handling the "field does not exist"
as a heuristic for "ah, because its not on this subtype".

---

This PR also happens to fix an unrelated bug with o2m collection
loading, which was introduced when we started allowing setting m2os
to just an id; i.e.:

em.create(Author, { publisher: "p:1" });