fix: Fix hard deletes showing up in toMatchEntity.
The recent fix of "finding m2os pointing to an id before the instance
was loaded" also picked up _deleted_ entities.

Which is technically fine b/c `.get` filters both hard and soft
deleted.

But it was a breaking change to `toMatchEntity`, which wanted to be
helpful and include soft-deleted entities in its assertions, but
was also picking up hard-deleted, which we do actually want to ignore.