fix: Fix performance issue into o2m.addedBeforeLoaded handling.
We recently fixed an issue with `author.books.get` missing books
that had been set to the author, before the author itself was
loaded from the db.

The fix was correct, but used a naive "scan all the things"
implementation which really does not perform well in large UoWs,
like over 10k entities.

Note that merely "scanning all loaded entities" (and commenting out
the `.includes` n^2 issue) still results in too much of a performance
regression, so we'll have to find a non-scanning way of fixing this
bug.

The original "author missing a book" bug has existed for a long time
and doesn't seem to have caused us any issues, so shouldn't be a lot
of harm in going back to living with it.