feat: Use CTE instead of UNIONs to batch queries.
We also know only batch structurally exact queries, which means
all queries that batched together will touch the same tables, have
the came join conditions, etc., they only differ in the parameters
in the WHERE clauses.

This capability is faciliated by the ParsedFindQuery, which gives us
an easy-to-work with AST that we can drop values from, then
JSON.stringify it, to group together similar-structure-differnet-value
queries.