fix: Transitively load derived values when the entities are new.
This prevents 'scopeUpdated has not been derived yet' errors, when
we've created a new entity, and some dependent field depends on a
field within the new entity.

Note that we're careful to only call `.load()` on derived fields
that are explicitly new/not set yet, as otherwise we'd defeat the
purpose of "derived fields can depend on other derived fields to
cheaply calculate their values".