fix: Restore populate hints cause loaded/live calcs.
A previous fix allowed derived fields to .load() their dependencies,
i.e. if those dependencies were themselves new and not yet calced.

However, it did that conditionally, i.e. if the dependency was already
calced, it would not re-calc it. Which is the desired behavior, to avoid
re-calculating a whole tree of derived values unnecessary.

That said, it broke that explicit `.populate` calls are the programmer
requesting a fresh calc, i.e. where they know they've recently mutated
the graph, so this PR restores the behavior.