fix: Avoid creating dups if different upsert clauses.
Previously multiple `findOrCreate`s with the same `where` but different
`ifNew`/`upsert` values were correctly getting batched into the same
`load(keys)` call, but we were still em.creating an entity for each
different ifNew/upsert clause.

Instead the intent was to just pick an arbitrary upsert clause as
a winner, which this PR actually does.

We could potentially also fail here, and let the user know they're
asking for the same entity to be set to different values; maybe as
an opt to findOrCreate.