fix: Fix findOrCreate with undefined. (#896)
The API for findOrCreate is opts-based, so wants `publisher: undefined`,
which matches Joist's opinionated choice of `undefined` to represent
empty within the object graph.

But when we call `em.find`, it drops `publisher: undefined` due to a
legacy-but-valid desire to match the behavior of `em.findGql` where
it's very easy for a resolver implementation that handles "the client
didn't include a filter for publisherId" to accidentally include the
`publisher` key on the `em.find` call.

I.e. something like:

```
const { publisherId } = filter;
await em.find(..., { publisherId });
```

Is the destructuring/restructuring idiom that looks innocent but will
mean the `publisherId` key is always set on the `em.find` call,
regardless of being included in the incoming filter.

Anyway, this watches for opts-find-undefined and converts over to
null so that the semantics work as expected.