feat: Add em.recalc, simplify em.touch.
Previously, em.touch did four things:

1. Bump updatedAt (if available)
2. Run hooks
3. Recalc all reactive derived fields
4. Revalidate all reactive rules

We have an issue in production where we're using em.touch to bump
updatedAt's up a tree (i.e. child changes, bump parent) for history
purposes.

However, the parent has an unrelated validation rule that validly
fails, i.e. the parent entity was valid in the past, but unrelated
changes have made one of its rules become invalid (because we
purposefully side-stepped the rule's reactivity).

(Specifically, the parent entity is a PRPC that now has an invalid
lot configuration, which is fine, but the configuration should only
be fixed the next time the users change/update the config, and not
part of the unrelated flow that just happens to be bumping the
parent.)

In addition to this specific production issue, we've aleady been wanting
to split up `em.touch` to add a `em.recalc` that specifically
recalculates fields and _doesn't_ bump `updatedAt` unless any of the
fields actually changed.

And for #4, trigger revalidate on all reactive rules, we currently don't
have an active use case for this; we're fine with the reactive rules
running organically as em.touch touches updatedAt or em.recalc changes
derived fields.

So, this PR:

* Changes `em.touch` to only bump `updatedAt` + run hooks
* Adds a new `em.touch` that only recalculates reactive fields (plus any
  downstream reactive fields if/as necessary)

This is a large/breaking change to the semantics of `em.touch`, but we
think it's warranted due as it more intuitively matches what "touch"
entails + also matches the Rails equivalent of "touch".