feat: Add preventEqualsOnEntities method for Jest.
Using toEquals with entities can cause issues b/c:

* The equality does a deep comparison of the entities & ends up diffing
  the entire EntityManager/ConnectionPool
* If equality fails, the toEquals diff similarily prints the entire
  entity + EntityManager + ConnectionPool

While it's possible to add custom equality matchers, to recognize the
entities and do a custom equals check, it's not possible to, if the
entities are _not_ equal, provide a custom diff output.

Because of this, we've created our own toMatchEntity, but it's hard
to remember to use it, so this prompts the prompt by just breaking
toEquals on entities all together.