fix: Restore soft-deleting filters on CTI base types.
We don't support CTI filtering on sub types, which should be easy to
make sure the alias is always included in the query, but just not
implemented yet.

The previous check I added for "we don't implement this yet" was too
strict though, as if we're querying against the base type directly
then we can do soft-delete filtering.