fix: Fix polys being too strict on the subtype matching.
If a poly pointed to a subtype (i.e. `task_old`), it's
`comp.otherMeta()` would correctly return `TaskOld`.

However the cstr from the `t:1` id would be just `Task`.

So the poly comp would think "this id is not mine", and return null.

So, we want to check the id cstr against the base meta...

...unless this poly does actually have different subtypes from the
same base type as separate components/columns, which seems a little
odd, but we have a test case covering, so it must have come up in
the past.