fix: Fix reacting to subtype-only relations.
If we saw a hint like:

* child (no subtypes) -> parent (that is a subtype) -> something else

When we would reverse it, and try to go:

* something else -> parent (that is a different subtype) -> children

The `parent[children]` would blow up, b/c that o2m may not exist.

This is very similar to conditionally walking through polys, so we
reuse the same `parent@Subtype` to infra to stop the walking before
trying to access `parent[children]`.