fix: Add tests for previous reactive hint fix, with more fixes.
This adds a test to reproduce the prior "walk through a different
subtype that does not have the relation" error, as well as applies
the fix to m2o and m2m handling as well.

There is also a related fix that runtime filtering of the type filter,
i.e. `publisher@Publisher` should pass for subtypes of `Publisher` like
either a `LargePublisher` or `SmallPublisher`, and previously we were
only exact-matching on the type name.