fix: Fix lens finds going through overwriteExisting path. (#1098)
The `EM.hydrate` method is a somewhat private-ish method for loading
an entity from a row result set.

It has conditional behavior on what to do if, when hydrating a row, that
row has already been loaded into an in-memory entity.

The safest option is to ignore the new hydrate, because it might right
over WIP changes in the current EM/entity.

However, the prior default was `overwriteExisting: true`, and then we
just put `overwriteExisting: false` in a bunch of places. Which was
fine, except we forgot one, in the lensDataLoader.

This surfaced as a bug in production not because of the incorrect overwriteExisting
behavior, but because it highlighted that the overwriteExisting behavior was really
pretty slow, because it had not been optimized to use the lazy row/data
serde hand-off.

This PR flips the default of hydrate to `overwriteExisting: false`,
which technically fixes the perf bug by skipping any hydrate logic for
rows that are already loaded, but we do go ahead and optimize the
`overwriteExisting: true` codepath as well.