fix: Add STI subtypes as per-type filters.
When we have base types, the normal filter only exposes the fields
available on the base type/all subtypes.

However we recently allowed to subtype-specific filtering with a
syntax of `<field><subtype>`, typed to the subtype filter, which
then allowed using each of the subtype's fields.

We built this for CTI, but STI should have gotten it for free, except
that the STI.subtypes field was not getting wired up correctly,
and so it just got skipped by accident.