fix: Fix RQF/RF dependencies.
This PR initially just added a notNull to `authors.favorite_book_id`,
to reproduce an issue with ReactiveReferences that are o2os.

However that change highlighted a bug that our RQF/RF loop was not
as tight as it should have been, so adds an explicit "run RQF first,
then RF second" in our em.flush RQF path.

commit-id:9b2c7911