feat: Recognize cross-field setDefault dependencies.
We use the field-level hints to async setDefault calls to topo
sort them.

Once we know the dependencies, we then call setDefaults in "Waves"
i.e. not strictly one-after-the-one, so that we still get parallelism
for fields that don't depend on each other.

We could further optimize this to do all hint loads in parallelize,
and only actually do the invocations synchronously in order.

Fixes #1160