fix: Allow returning Reacted entities from hasReactiveAsyncProperty.
Previously the Reacted<Author> was expected to be "the full Author",
which of course it's not, given that Reacted are subviews of the full
type.

This PR adjusts the return type of hasReactiveAsyncProperty to, if
either an entity or optional entity (i.e. `entity | undefined`) is
returned, we'll accept the reacted subtype.

We piggyback on the existing `__orm` type hint field, albeit we didn't
have the entityType itself in the type before, so we codegen that out
now.