feat: Add EntityManager.mode.
Allows putting the EM into either read-only (no mutations allowed) or
in-memory-writes (mutations allowed but no writes).

The in-memory-writes in particular is useful for building "oracle"
functionality where you want to mutate entities to "see what would happen",
but ensure the WIP data is never committed to the db, even by an errant,
over-eager `em.flush`.