feat: Update factories to set defaults immediately.
Because factory entities are DeepNew/deeply loaded, it's very likely
that what would, in production, usually be an async `setDefault` call
(requiring a Promise to load data from the db) can, in a test, actually
be done synchronously (by assuming the test instance has the data
already in memory).

This PR leans into that assumption, and uses a cute `async
() => {}.constructor` check to guess at whether a `setDefault`
lambda is synchronous or not.

It's not foolproof but should work.