feat: Add updatePartial method.
Currently em.createOrUpdatePartial requires *not* passing an entity,
i.e. a constructor + opts (with id) combo.

This PR adds updatePartial, which is like `createOrUpdatePartial` in
that it takes a deep tree of data, but also like `setPartial` in that
it works against an existing entity.

I'm currently just exposing this as a static function, but it could
maybe be a method on `EntityManager` (since it's async) or maybe
the `Entity` itself, like maybe `setPartialDeep` or `setDeepPartial`?