feat: Allow o2m subtype specialization.
Similar to a SmallPublisher.group: SmallPublisherGroup m2o
specialization, allow a SmallPublisherGroup.publishers: SmallPublishers
specialization.

Note that these may go hand-in-hand, and be set on both sides of a
relation, but not necessarily, i.e. it could be valid for the
SmallPublisherGroup to have any type of Publisher, until this hint
was specifically added, in addition to the "SmallPublisher.group is a
SmallPublisherGroup hint".