fix: Fix toString to sequentially assign new entities.
Previously the `1` part of a new entity's `Author#1` toString() was
"the index within all authors in the EntityManager".

This was fine, but I had been using the `#1` as a quick easy to
count the number of entities created during an em.flush, i.e.
seeing `Author#5` === "the 5th newly created Author".

But previously the `#5` could have been "the 1st newly created
Author, after we loaded 4 other existing Authors".

This PR changes the `#` numbers to only increment/indexOf against
other also-newly-created entities.