fix: Don't mutate between filters.
When copying a user-provided `where` clause into a `ParsedFindQuery`,
we create an AST of `{ kind: "between", value: ...user's array...} `.

This is fine, but mapToDb was directly mutating the array's values
from domain types to db types, i.e. when doing Temporal ZonedDateTime
conversion.

This fix just makes sure that we assign `filter.value` to a newly
created array, which matches how the rest of the `mapDb` / `filter`
rewriting works.