feat: Allow factories to set derived values. (#1344)
* feat: Allow factories to set derived values.

Normally business logic is prevented from setting ReactiveFields
and ReactiveQueryFields.

However, for unit tests, it can be useful to set a RF/RFQ to a specific
value, as part of test setup, without fully re-creating the data
necessary to get that value calculated by the RF/RFQ logic.

So now any RF/RFQ will have a `with${fieldName}` factory option
that, when provided, will be used/get stored in the db, instead of
the RF/RFQs canonical operation.

Note that this only applies to the factory-created instance, i.e.
if code-under-test happens to load the RF/RFQ on a new entity, and
cause it to be re-calculated, then it will be recalculated, i.e.
we don't "remember" that it should have a pinned value.

* Remove as.