fix: Better error message for unsupported reactivity. (#1356)
Currently we don't support reacting to ReactiveReferences, because they
don't have an in-memory "other side" of their relation.

I.e. with a ReactiveReference of Author.favoriteBook, a rule like:

```ts
authorConfig.addRule({ favoriteBook: "title", firstName: {} }, a => {
  return a.firstName === a.favoriteBook.get.title ? "bad" : undefined;
});
```

When `Book.title` changes, our reaction logic would try to:

"Look for author's that I'm the favorite book of", which would be
the relation `book.favoritedByAuthors` and "follow" that relation
to say "tell those authors to rerun their validation rule".

This reactivity leverages the "other side" relation, i.e.
`book.favoritedByAuthors`, however we currently do not codegen
the "other side" collection for ReactiveReferences because it's
difficult to keep "in-sync" (which we do for o2m relations) because:

a) there could be many different writes that invalidate a RR and
need "the other side" to be recalcd, and

b) the recalc would likely need to be async

We might be able to solve this someday, but for now just error out with
a better message.