fix: Fix setting over ReactiveReferences.
Calling `em.create(Author, { favoriteBook: b1 })` would work because
invoking `setOpts` from the constructor meant that `a.favoriteBook`
was not initialized yet, so the `current instanceof ...` logic
in `setOpts` was not kicking in.

Now, we have `em.create` not call `setOpts` until after the instance
is fully initialized.

We don't go all the way to hard deprecating the `new Author(em, ...)`
syntax, but we should plan on doing so.

Fixes #1442.