chore: disallow `this` in property callbacks (#1595)
Due to the changes in #1586, this inside a property initialization callback (such as for hasAsyncProp) no longer refers to the instance being instantiated, but rather to the fake instance generated by getProperties. This can result in unpredictable behavior and is tricky to debug, as often the error won't reveal itself immediately upon access. This PR addresses this by making any access of a property on this inside one of these callbacks immediately throw an error. We accomplish this by deleting the "own" properties on the fake instance and then installing a proxy as its prototype, which allows us to respond to any access by immediately failing.