feat: Trim string values. (#1664)
This builds on our existing "sanitize empty string to undefined", and
just extends it to leading/trailing spaces.

This is automatically disabled if the varchar's default value is `""`,
which is how the varchar can declare "i'm okay with empty strings"
(obviously, since it has that has its default value).

We could also add sanitize knobs to `joist-config.json` if/as needed in
the future.